+++ CHANGELOG.md
@@ -5,10 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.0] - 2025-09-17
 
 ### Added
-- Your next feature here...
+- New feature to control compression quality for JPEG and WEBP formats.
+- Added a `--quality` command-line argument (1-100) to specify the compression level.
+
+### Changed
+- Enhanced the `--help` message to include a section with recommended settings and examples for common use cases.
 
 ## [0.4.0] - 2025-09-17